server/requirements.txt
fastapi
uvicorn
python-dotenv
requests
loguru